+++ platform/jvm/capture/src/main/kotlin/io/bitdrift/capture/CaptureJniLibrary.kt
@@ -185,11 +185,10 @@ internal object CaptureJniLibrary : IBridge, IStreamingReportProcessor {
      * @param fields fields map of arbitrary key-value fields to include with the log.
      * @param matchingFields fields map of arbitrary key-value fields that can be read when processing
      *        a given log but are not a part of the log itself.
-     * @param overrideExpectedPreviousProcessSessionId if provided, it will be assigned as a session ID
-     *        of the log. The passed value has to match the last session ID active during the previous run of
-     *        the SDK.
-     * @param overrideOccurredAtUnixMilliseconds used to override the timestamp of the log. Used only if
-     *        valid `override_expected_previous_process_session_id` is provided, ignored otherwise.
+     * @param usePreviousProcessSessionId if set to true, this log will be emitted with the session ID
+     *        corresponding to the last session ID during the previous process run. This is helpful for reporting
+     *        crashes on next app launch.
+     * @param overrideOccurredAtUnixMilliseconds used to override the timestamp of the log.
      * @param blocking if true, the call blocks until the log has been processed.
      */
     external fun writeLog(
@@ -199,10 +198,16 @@ internal object CaptureJniLibrary : IBridge, IStreamingReportProcessor {
         log: String,
         fields: Map<String, FieldValue>,
         matchingFields: Map<String, FieldValue>,
-        overrideExpectedPreviousProcessSessionId: String?,
+        usePreviousProcessSessionId: Boolean,
         overrideOccurredAtUnixMilliseconds: Long,
         blocking: Boolean,
     )
+
+    /**
+     * Shuts down the logger, blocking until the event loop has terminated. This is not yet ready
+     * to be exposed as a public API due to lack of testing and no timeout on the blocking wait.
+     */
+    external fun shutdown(loggerId: Long)
 
     /**
      * Writes a session replay log.