+++ platform/jvm/capture/src/main/kotlin/io/bitdrift/capture/events/lifecycle/AppExitLogger.kt
@@ -20,7 +20,6 @@ import io.bitdrift.capture.LogLevel
 import io.bitdrift.capture.LogType
 import io.bitdrift.capture.LoggerImpl
 import io.bitdrift.capture.common.ErrorHandler
-import io.bitdrift.capture.common.IBackgroundThreadHandler
 import io.bitdrift.capture.common.Runtime
 import io.bitdrift.capture.common.RuntimeFeature
 import io.bitdrift.capture.events.performance.IMemoryMetricsProvider
@@ -33,9 +32,7 @@ import io.bitdrift.capture.reports.exitinfo.LatestAppExitReasonResult
 import io.bitdrift.capture.reports.jvmcrash.CaptureUncaughtExceptionHandler
 import io.bitdrift.capture.reports.jvmcrash.ICaptureUncaughtExceptionHandler
 import io.bitdrift.capture.reports.jvmcrash.IJvmCrashListener
-import io.bitdrift.capture.threading.CaptureDispatchers
 import io.bitdrift.capture.utils.BuildVersionChecker
-import java.nio.charset.StandardCharsets
 
 internal class AppExitLogger(
     private val logger: LoggerImpl,
@@ -43,7 +40,6 @@ internal class AppExitLogger(
     private val runtime: Runtime,
     private val errorHandler: ErrorHandler,
     private val versionChecker: BuildVersionChecker = BuildVersionChecker(),
-    private val backgroundThreadHandler: IBackgroundThreadHandler = CaptureDispatchers.CommonBackground,
     private val memoryMetricsProvider: IMemoryMetricsProvider,
     private val latestAppExitInfoProvider: ILatestAppExitInfoProvider = LatestAppExitInfoProvider,
     private val captureUncaughtExceptionHandler: ICaptureUncaughtExceptionHandler = CaptureUncaughtExceptionHandler,
@@ -71,29 +67,10 @@ internal class AppExitLogger(
         }
         captureUncaughtExceptionHandler.install(this)
         logPreviousExitReasonIfAny()
-        backgroundThreadHandler.runAsync {
-            saveCurrentSessionId()
-        }
     }
 
     fun uninstallAppExitLogger() {
         captureUncaughtExceptionHandler.uninstall()
-    }
-
-    @RequiresApi(Build.VERSION_CODES.R)
-    fun saveCurrentSessionId(sessionId: String? = null) {
-        if (!runtime.isEnabled(RuntimeFeature.APP_EXIT_EVENTS) || !versionChecker.isAtLeast(Build.VERSION_CODES.R)) {
-            return
-        }
-
-        val currentSessionId = sessionId ?: logger.sessionId
-
-        try {
-            activityManager.setProcessStateSummary(currentSessionId.toByteArray(StandardCharsets.UTF_8))
-        } catch (error: Throwable) {
-            // excessive calls to this API could result in a RuntimeException.
-            errorHandler.handleError("Failed to save session id in ActivityManager", error)
-        }
     }
 
     @RequiresApi(Build.VERSION_CODES.R)
@@ -109,22 +86,13 @@ internal class AppExitLogger(
 
             is LatestAppExitReasonResult.Valid -> {
                 val lastExitInfo = lastExitInfoResult.applicationExitInfo
-                // extract stored id from previous session in order to override the log,
-                // bail and report error if not present
-                val sessionId =
-                    lastExitInfo.processStateSummary?.toString(StandardCharsets.UTF_8)
-                if (sessionId == null) {
-                    val errorMessage = "AppExitLogger: processStateSummary from ${lastExitInfo.processName} is null."
-                    errorHandler.handleError(errorMessage)
-                    return
-                }
 
                 val timestampMs = lastExitInfo.timestamp
                 logger.log(
                     LogType.LIFECYCLE,
                     lastExitInfo.reason.toLogLevel(),
                     buildAppExitInternalFieldsMap(lastExitInfo),
-                    attributesOverrides = LogAttributesOverrides.SessionID(sessionId, timestampMs),
+                    attributesOverrides = LogAttributesOverrides.PreviousRunSessionId(timestampMs),
                 ) { APP_EXIT_EVENT_NAME }
             }
         }