+++ platform/jvm/capture/src/main/kotlin/io/bitdrift/capture/reports/processor/IssueReporterProcessor.kt
@@ -7,7 +7,10 @@
 
 package io.bitdrift.capture.reports.processor
 
+import android.os.Build
+import android.os.strictmode.Violation
 import android.util.Log
+import androidx.annotation.RequiresApi
 import com.google.flatbuffers.FlatBufferBuilder
 import io.bitdrift.capture.BuildConstants
 import io.bitdrift.capture.Capture.LOG_TAG
@@ -141,17 +144,49 @@ internal class IssueReporterProcessor(
      * NOTE: This will need to run by default on the caller thread
      */
     fun persistJvmCrash(
-        timestamp: Long,
         callerThread: Thread,
         throwable: Throwable,
         allThreads: Map<Thread, Array<StackTraceElement>>?,
     ) {
+        persistJvmIssue(
+            callerThread = callerThread,
+            throwable = throwable,
+            allThreads = allThreads,
+            reportType = ReportType.JVMCrash,
+            isFatal = true,
+        )
+    }
+
+    /**
+     * Process StrictMode violations into a packed format
+     *
+     * NOTE: This will need to run by default on the caller thread
+     */
+    @RequiresApi(Build.VERSION_CODES.P)
+    fun persistStrictModeViolation(violation: Violation) {
+        persistJvmIssue(
+            callerThread = Thread.currentThread(),
+            throwable = violation,
+            allThreads = Thread.getAllStackTraces(),
+            reportType = ReportType.StrictModeViolation,
+            isFatal = false,
+        )
+    }
+
+    private fun persistJvmIssue(
+        callerThread: Thread,
+        throwable: Throwable,
+        allThreads: Map<Thread, Array<StackTraceElement>>?,
+        reportType: Byte,
+        isFatal: Boolean,
+    ) {
+        val timestamp = dateProvider.invoke().time
         val builder = FlatBufferBuilder(FBS_BUILDER_DEFAULT_SIZE)
         val sdk = createSDKInfo(builder)
         val appMetrics = createAppMetrics(builder)
         val deviceMetrics = createDeviceMetrics(builder, timestamp)
         val report =
-            JvmCrashProcessor.getJvmCrashReport(
+            JvmProcessor.getJvmReport(
                 builder,
                 sdk,
                 appMetrics,
@@ -159,14 +194,23 @@ internal class IssueReporterProcessor(
                 throwable,
                 callerThread,
                 allThreads,
+                reportType,
             )
         builder.finish(report)
 
-        reporterIssueStorage.persistFatalIssue(
-            timestamp,
-            builder.sizedByteArray(),
-            ReportType.JVMCrash,
-        )
+        if (isFatal) {
+            reporterIssueStorage.persistFatalIssue(
+                timestamp,
+                builder.sizedByteArray(),
+                reportType,
+            )
+        } else {
+            reporterIssueStorage.persistNonFatalIssue(
+                timestamp,
+                builder.sizedByteArray(),
+                reportType,
+            )
+        }
     }
 
     private fun createSDKInfo(builder: FlatBufferBuilder): Int =